check calibration update for noetic
These changes make the following code run in ROS Noetic
with Ubuntu 20.04 and Python 3. However, this code has
not been fully tested. The robot moves and appears to
collect data. A fit error appears to be calculated.
The visualization of the fit may have a problem.

+ stretch_calibration collect_check_head_calibration_data.launch
+ stretch_calibration check_head_calibration.sh

For now, I've commented out "class HelloNode" in hello_misc.py
due to a syntax error without an obvious fix.