Add HelloNode.get_joint_state() + other improvements
The get_joint_state() method makes it easier to get
information about a joint, including whether the
joint is moving. A mode variable is added to make
it easier to determine what mode the driver is in.
Lastly, some of HelloNode's variables are made private
to make it easier to use the class from iPython.